refactor structure and couple of fixes
- create separate trait for collection specific methods
- fix downloads to custom path
- add connection checks and catches
